SERVER-17822 Make OpDebug::writeConflicts a 64-bit counter

+++ src/mongo/db/curop.h
@@ -172,7 +172,7 @@ namespace mongo {
         bool fastmodinsert;  // upsert of an $operation. builds a default object
         bool upsert;         // true if the update actually did an insert
         int keyUpdates;
-        int writeConflicts;
+        long long writeConflicts;
         ThreadSafeString planSummary; // a brief std::string describing the query solution
 
         // New Query Framework debugging/profiling info
